Moves the search for user feature to the proper class.
