+++ core/src/main/kotlin/chat/rocket/core/internal/rest/ChatRoom.kt
@@ -43,13 +43,9 @@ suspend fun RocketChatClient.getMembers(
         RestResult::class.java,
         Types.newParameterizedType(List::class.java, User::class.java)
     )
-    val result = handleRestCall<RestResult<List<User>>>(request, type)
 
-    return@withContext PagedResult<List<User>>(
-        result.result(),
-        result.total() ?: 0,
-        result.offset() ?: 0
-    )
+    val result = handleRestCall<RestResult<List<User>>>(request, type)
+    return@withContext PagedResult<List<User>>(result.result(), result.total() ?: 0, result.offset() ?: 0)
 }
 
 /**
@@ -79,13 +75,9 @@ suspend fun RocketChatClient.getFavoriteMessages(
         RestResult::class.java,
         Types.newParameterizedType(List::class.java, Message::class.java)
     )
-    val result = handleRestCall<RestResult<List<Message>>>(request, type)
 
-    return@withContext PagedResult<List<Message>>(
-        result.result(),
-        result.total() ?: 0,
-        result.offset() ?: 0
-    )
+    val result = handleRestCall<RestResult<List<Message>>>(request, type)
+    return@withContext PagedResult<List<Message>>(result.result(), result.total() ?: 0, result.offset() ?: 0)
 }
 
 /**
@@ -116,13 +108,9 @@ suspend fun RocketChatClient.getPinnedMessages(
         RestResult::class.java,
         Types.newParameterizedType(List::class.java, Message::class.java)
     )
-    val result = handleRestCall<RestResult<List<Message>>>(request, type)
 
-    return@withContext PagedResult<List<Message>>(
-        result.result(),
-        result.total() ?: 0,
-        result.offset() ?: 0
-    )
+    val result = handleRestCall<RestResult<List<Message>>>(request, type)
+    return@withContext PagedResult<List<Message>>(result.result(), result.total() ?: 0, result.offset() ?: 0)
 }
 
 /**
@@ -153,8 +141,8 @@ suspend fun RocketChatClient.getFiles(
         RestResult::class.java,
         Types.newParameterizedType(List::class.java, GenericAttachment::class.java)
     )
-    val result = handleRestCall<RestResult<List<GenericAttachment>>>(request, type)
 
+    val result = handleRestCall<RestResult<List<GenericAttachment>>>(request, type)
     return@withContext PagedResult<List<GenericAttachment>>(result.result(), result.total() ?: 0, result.offset() ?: 0)
 }
 
@@ -190,38 +178,6 @@ suspend fun RocketChatClient.joinChat(roomId: String): Boolean = withContext(Com
     val request = requestBuilder(url).post(body).build()
 
     return@withContext handleRestCall<BaseResult>(request, BaseResult::class.java).success
-}
-
-/**
- * Returns the list of user of a chat room that satisfies a query.
- *
- * @param queryParam Parameter which is used to query users on the basis of regex.
- * @param count The number of users to be returned in the result
- * @param offset The number of users to skip from the beginning
- * @return The list of user of a chat room that satisfies a query.
- */
-suspend fun RocketChatClient.queryUsers(
-    queryParam: String,
-    count: Long = 30,
-    offset: Long = 0
-): PagedResult<List<User>> = withContext(CommonPool) {
-    val httpUrl = requestUrl(restUrl, "users.list")
-        .addQueryParameter("query", "{ \"name\": { \"\\u0024regex\": \"$queryParam\" } }")
-        .addQueryParameter("offset", offset.toString())
-        .addQueryParameter("count", count.toString())
-        .build()
-    val request = requestBuilder(httpUrl).get().build()
-    val type = Types.newParameterizedType(
-        RestResult::class.java,
-        Types.newParameterizedType(List::class.java, User::class.java)
-    )
-
-    val result = handleRestCall<RestResult<List<User>>>(request, type)
-    return@withContext PagedResult<List<User>>(
-        result.result(),
-        result.total() ?: 0,
-        result.offset() ?: 0
-    )
 }
 
 /**
